Fixing a bug when changing the orientation of the snake
So, I had a problem with input lag using the pygame tools, so, to fix this I wait 2 iterations of the main loop to change the state of the current board, to its next state, the problem was the next. Imagine u got a snake with length 2 or more, in a position like this:  □□□□ --->, this snake is going to the right, the posible moves, is keep going foward, turn right, and turn left, and you cant go backwards because you kill youself, when I change the orientation of the head I check the current facing state and then change the orientation if it's posible. The problem comes when I take 2 iterations of the main loop to move the snake, cause the user has time to change the orientation 2 times, so, if the snake was facing East, like this □□□□ --->, and the user press, down, and left quickly enough, instead of moving down, and then left, the program'd change the orientation down, then left, cause it's already facing down, but hasn't move yet, then the state'd be the next   □□□□ <-----, so the head is looking to it's body, to fix this, I save the current facing, and every time I try to change the orientation I check the current facing state. Also created nextFacing, so in the move function facing <---- nextFacing